+++ src/main/groovy/com/github/best/champion/region/season_4/fetch_recent_games_from_ranked_summoners.groovy
@@ -6,7 +6,7 @@ import com.mongodb.BasicDBObject
 import groovy.time.TimeDuration
 import groovyx.net.http.HttpResponseException
 
-def sinceHours = args.length > 1 ? args[0] : 6
+def sinceHours = args.length > 1 ? args[0] as int : 6
 timeDuration = new TimeDuration(sinceHours, 0, 0, 0)
 
 MongoUtils.connect {
